export default garden item to file
